+++ src/main/java/com/application/close/mqtt/payload/MqttConnectResp.java
@@ -1,0 +1,20 @@
+package com.application.close.mqtt.payload;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
+public class MqttConnectResp {
+
+	private String clientId;
+
+	private String mqttUrl;
+
+	private boolean isConnected;
+
+}