+++ src/main/java/com/application/close/mqtt/payload/MqttParamPayload.java
@@ -34,11 +34,11 @@ public class MqttParamPayload {
 	@Max(value = 65535, message = "Port must be <= 65535")
 	private String port;
 
-	@NotBlank(message = "Username is required")
+//	@NotBlank(message = "Username is required")
 	@Size(min = 3, max = 10, message = "Username must be between 3 and 10 characters")
 	private String userName;
 
-	@NotBlank(message = "Password is required")
+//	@NotBlank(message = "Password is required")
 	@Size(min = 8, max = 15, message = "Password must be between 8 and 15 characters")
 	private String password;