+++ src/main/java/com/application/close/mqtt/service/MqttService.java
@@ -1,0 +1,11 @@
+package com.application.close.mqtt.service;
+
+import com.application.close.mqtt.payload.MqttConnectResp;
+
+public interface MqttService {
+
+	MqttConnectResp connect(int mqttParamId);
+
+	MqttConnectResp connectWithUserName(int mqttParamId);
+
+}